:sparkles: [#155] Add shipping cost model to shop configuration
With the shipping cost model, it's possible to define weight ranges and
their associated costs, for each supported country.

The admin allows configuration of these ranges.

Additionally, the order model now tracks the delivery method - shipping
by mail or picking up in the store itself. We also track the resulting
shipping costs on the order model, making it part of the snapshot
data when an order is placed.