:recycle: #155 -- Rework the validation error introspection code
Due to the type definitions we can check specific fields to derive
whether there are validation errors or not.

The validation state is moved into the context so that the redirect
utility can use it while being a prop-less component to hook up to
data routers, and the nav controls make use of the same error
information.

Later on, we can hook up the validation errors to the initialErrors of
the formik state too.