:bug: Ensure slug is exported with products
Not exporting the slug (or populating it) leads to database errors. The
easiest is to export existing slugs.

Manufacturer has been dropped from the export format for performance
reasons, the primary keys don't say anything meaningful anyway. This
can be managed via the admin interface.

The category field is dropped, because it doesn't exist. At some point,
this changed into a m2m categories field, which is also better managed
via the admin interface.