:green_heart: Fix CI failure
The amount of categories created rolled over to sequence 10 when
the last category/second root child is being created, resulting
in a lexical sort putting that second child *before* the first
child which is category-6 or so, and that messes up the tree
order because the model defines an ordering-by-field property
of name.

Fixed by adding leading zeros to the name field from the
factory that should be good until 9999 instances.